requirements.txt
## RAG pipeline dependencies - Haystack
scikit-learn
openai>=1.0.0
requests>=2.31.0
chromadb>=0.5.0
haystack-ai>=2.0.0
chroma-haystack>=0.18.0
sentence-transformers>=2.7.0
nltk>=3.9.1
PyYAML>=6.0.0
tiktoken
